Added a logging class
Created a logging class that allows logs to be sent to (initially) the
console. But can be adapted to allow other logging services. Currently
it can handle Notices, Warnings and Errors and features appropriate
color codes when outputting to the CLI.

It can be called by using Log.error(message), Log.warning(message) or
Log.notice(message)